Flattened create command
